backend/payments/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
asyncpg
psycopg2-binary
pytest
httpx
redis
rq
python-dotenv
debugpy
